+++ package.json
@@ -61,7 +61,7 @@
     "deepmerge": "^4.3.1",
     "dropbox": "git+https://lovegaoshi@github.com/lovegaoshi/dropbox-sdk-js.git",
     "event-target-polyfill": "^0.0.4",
-    "expo": "^52.0.25",
+    "expo": "^52.0.27",
     "expo-auth-session": "~6.0.2",
     "expo-clipboard": "~7.0.1",
     "expo-crypto": "~14.0.2",
@@ -78,7 +78,7 @@
     "js-base64": "^3.7.7",
     "libmuse": "https://github.com/lovegaoshi/muse.git#commit=942f16bde7ba44ad1d0452ce838a96b17fe95aff",
     "lodash": "^4.17.21",
-    "lottie-react-native": "7.2.1",
+    "lottie-react-native": "7.2.2",
     "lru-cache": "7.18.3",
     "md5": "^2.3.0",
     "metro": "^0.81.0",
@@ -99,7 +99,7 @@
     "react-native-get-random-values": "^1.11.0",
     "react-native-image-colors": "^2.4.0",
     "react-native-lyric": "https://lovegaoshi@github.com/lovegaoshi/react-native-lyric.git#commit=6f20e83948c29b0d46833ab9173cd81f99d0ab48",
-    "react-native-pager-view": "6.6.1",
+    "react-native-pager-view": "6.7.0",
     "react-native-paper": "^5.13.1",
     "react-native-purchases": "^8.5.0",
     "react-native-qrcode-svg": "^6.3.14",
@@ -121,7 +121,7 @@
     "use-debounce": "^10.0.4",
     "uuid": "^11.0.5",
     "web-streams-polyfill": "^4.1.0",
-    "youtubei.js": "^12.2.0",
+    "youtubei.js": "^13.0.0",
     "zustand": "^5.0.3"
   },
   "resolutions": {
@@ -144,7 +144,7 @@
     "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.14",
     "@types/md5": "^2.3.5",
-    "@types/node": "^22.10.6",
+    "@types/node": "^22.10.7",
     "@types/qs": "^6.9.18",
     "@types/react": "~18.3.18",
     "@types/react-native": "^0.73.0",
@@ -152,9 +152,9 @@
     "@types/react-native-share-menu": "^5.0.5",
     "@types/react-test-renderer": "^19.0.0",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.20.0",
-    "@typescript-eslint/parser": "^8.20.0",
-    "@welldone-software/why-did-you-render": "^8.0.3",
+    "@typescript-eslint/eslint-plugin": "^8.21.0",
+    "@typescript-eslint/parser": "^8.21.0",
+    "@welldone-software/why-did-you-render": "^10.0.1",
     "argparse": "^2.0.1",
     "babel-jest": "^29.7.0",
     "babel-plugin-module-resolver": "^5.0.2",
@@ -167,7 +167,7 @@
     "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.2.2",
+    "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.1.0",
     "gts": "^6.0.2",
