+++ patches/youtubei.js+13.0.0.patch
